+++ docs/executor/legacy_balsam_executor.rst
@@ -1,16 +1,16 @@
-Balsam 1 MPI Executor
-=====================
+Legacy Balsam MPI Executor
+==========================
 
-.. automodule:: balsam_executor
+.. automodule:: legacy_balsam_executor
   :no-undoc-members:
 
-.. autoclass:: BalsamMPIExecutor
+.. autoclass:: LegacyBalsamMPIExecutor
   :show-inheritance:
   :inherited-members:
   :member-order: bysource
   :members: __init__, submit, poll, manager_poll, kill, set_kill_mode
 
-.. autoclass:: BalsamTask
+.. autoclass:: LegacyBalsamTask
   :show-inheritance:
   :member-order: bysource
   :members: workdir_exists, file_exists_in_workdir, read_file_in_workdir, stdout_exists, read_stdout
