+++ libensemble/tests/balsam_tests/reset_balsam_tests.py
@@ -2,8 +2,8 @@
 
 import balsam.launcher.dag as dag
 
-dag.BalsamTask.objects.filter(name__contains='outfile').delete()
+dag.LegacyBalsamTask.objects.filter(name__contains='outfile').delete()
 
-for job in dag.BalsamTask.objects.filter(name__contains='job_test_balsam'):
+for job in dag.LegacyBalsamTask.objects.filter(name__contains='job_test_balsam'):
     job.update_state('CREATED')
     job.save()
