+++ libensemble/tests/standalone_executor_tests/test_executor_manager_poll.py
@@ -40,8 +40,8 @@ USE_BALSAM = False  # Take as arg
 
 # Create and add exes to registry
 if USE_BALSAM:
-    from libensemble.executors.balsam_executor import BalsamMPIExecutor
-    exctr = BalsamMPIExecutor()
+    from libensemble.executors.balsam_executor import LegacyBalsamMPIExecutor
+    exctr = LegacyBalsamMPIExecutor()
 else:
     from libensemble.executors.mpi_executor import MPIExecutor
     exctr = MPIExecutor()
