+++ libensemble/tests/standalone_executor_tests/test_executor_multi.py
@@ -34,8 +34,8 @@ USE_BALSAM = False  # Take as arg
 
 # Create and add exes to registry
 if USE_BALSAM:
-    from libensemble.baslam_executor import BalsamMPIExecutor
-    exctr = BalsamMPIExecutor()
+    from libensemble.baslam_executor import LegacyBalsamMPIExecutor
+    exctr = LegacyBalsamMPIExecutor()
 else:
     from libensemble.executors.mpi_executor import MPIExecutor
     exctr = MPIExecutor()
